+++ src/i18n/locales/ca/create-[slug].ts
@@ -1,4 +1,4 @@
 export default {
   metadataDescription:
-    'Crea ràpidament el mem {name} en línia, sense marca d’aigua, de franc i totalment personalitzat.'
+    'Crea ràpidament el mem {name} en línia, de franc i totalment personalitzat.'
 } as const