+++ src/i18n/locales/en/create-[slug].ts
@@ -1,4 +1,4 @@
 export default {
   metadataDescription:
-    'Quickly create the meme {name} online without watermark, for free, and completely personalized.'
+    'Quickly create the meme {name} online, for free, and completely personalized.'
 } as const